requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
black==19.10b0
cfgv==3.1.0
click==7.1.1
colorama==0.4.3
coverage==5.0.4
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
identify==1.4.13
mccabe==0.6.1
more-itertools==8.2.0
nodeenv==1.3.5
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
pre-commit==2.2.0
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pyserial==3.4
pytest==5.4.1
pytest-cov==2.8.1
PyYAML==5.3.1
regex==2020.2.20
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
virtualenv==20.0.13
wcwidth==0.1.9
